pyproject.toml
[project]
name = "nycmesh-meshdb"
version = "0.1"
dependencies = [
    "celery[redis]==5.3.*",
    "django==4.2.*",
    # We need djangorestframework features that are coming in 3.15.*
    # but that release hasn't shipped yet, so grab the commit that is likely to ship.
    # This will likely have a small diff with the final 3.15.* release, but is much better
    # then pulling their master branch (which changes constantly)
    # FIXME: Go back to PyPI once https://github.com/encode/django-rest-framework/pull/8794 is merged
    "djangorestframework@git+https://github.com/encode/django-rest-framework.git@4bbfa8d4556b5847e91ba95f457cc862b7a0f027",
    "drf-hooks==0.1.3",
    "psycopg2-binary==2.9.*",
    "gunicorn==22.0.*",
    "python-dotenv==1.0.*",
    "stringcase==1.2.*",
    "python-dotenv==1.0.*",
    "py3-validate-email@git+https://gitea.ksol.io/karolyi/py3-validate-email#egg=@v1.0.9",
    "phonenumbers==8.13.*",
    "requests==2.32.*",
    "geopy==2.4.*",
    "django-cors-headers==4.3.*",
    "nameparser==1.1.*",
    "inflect==7.0.*",
    "fastkml[lxml]==1.a10", # Update me to stable, once 1.0 is released officially
    "drf-spectacular==0.27.*",
    "djangorestframework-dataclasses==1.3.*",
    "django-nonrelated-inlines==0.2.*",
    "django-filter==24.1",
    # FIXME: Go back to PyPI when https://github.com/bhch/django-jsonform/pull/162 is merged
    "django-jsonform@git+https://github.com/willnilges/django-jsonform.git@51cbed42ccdaec81a35c97a919d054e2c9ca0207",
    "faker==24.3.*",
    # FIXME: Go back to PyPI when https://github.com/jazzband/django-dbbackup/pull/515 or https://github.com/jazzband/django-dbbackup/pull/511 is merged
    "django-dbbackup@git+https://github.com/willnilges/django-dbbackup.git@62048411ff5beac4beac1578f686824214f1f33a",
    "django-storages==1.14.*",
    "django-stubs[compatible-mypy]==5.0.*",
    "djangorestframework-stubs[compatible-mypy]==3.15.*",
    "django-filter-stubs==0.1.*",
    "boto3==1.34.*",
    "six==1.16.0",
]

[project.optional-dependencies]
dev = [
    "invoke == 2.2.*",
    "pytest == 7.4.*",
    "flake8 == 6.1.*",
    "black == 23.7.*",
    "isort == 5.12.*",
    "coverage == 7.3.*",
    "mypy == 1.5.*",
    "flask == 3.0.*",
    "django-cprofile-middleware==1.0.5",
    "django-silk==5.1.0",
    "types-requests==2.32.*",
    "types-six==1.16.0.*",
]

[project.scripts]
meshdb-spreadsheet-import = "meshdb.utils.spreadsheet_import.main:main"

[build-system]
requires = [ "setuptools >= 61", "wheel", "mypy" ]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]


[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
check_untyped_defs = true
show_error_codes = true
pretty = true
ignore_missing_imports = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
warn_unused_configs = true
exclude = [
    "^src/meshdb/utils/spreadsheet_import/.*",
    "^src/meshapi/tests/.*",
    "^src/meshapi_hooks/tests/.*",
    "^src/meshapi/migrations/.*",
    "^src/meshapi/util/django_pglocks.py",
    "^src/meshapi/models/util/custom_many_to_many.py",
    "^src/meshapi/docs.py"
]

[tool.django-stubs]
django_settings_module = "meshdb.settings"


[tool.black]
line-length = 120
include = '\.pyi?$'


[tool.isort]
profile = "black"
line_length = 120
known_first_party = [ "meshdb" ]

